deployment/requirements.txt
streamlit==1.32.0
pandas==2.2.2
plotly==5.24.0
matplotlib==3.7.1
seaborn==0.12.2
numpy==1.26.4
Pillow==10.4.0
tensorflow==2.15.0
protobuf==3.20.3